naming and aggregator function cleanup
